+++ src/collective/polls/tests/test_setup.py
@@ -14,6 +14,7 @@ JS = (
     '++resource++collective.polls/js/jquery.flot.pie.js',
     '++resource++collective.polls/js/polls.js',
     '++resource++collective.polls/js/collective.poll.js',
+    '++resource++collective.polls/js/jquery.tasksplease.js',
 )
 
 CSS = (
@@ -35,7 +36,7 @@ class InstallTestCase(unittest.TestCase):
     def test_profile_version(self):
         setup_tool = self.portal['portal_setup']
         self.assertEqual(
-            setup_tool.getLastVersionForProfile(PROFILE), (u'5',))
+            setup_tool.getLastVersionForProfile(PROFILE), (u'6',))
 
     def test_add_permission(self):
         permission = 'collective.polls: Add poll'