+++ src/collective/polls/upgrades/__init__.py
@@ -3,15 +3,15 @@ from collective.polls.logger import logger
 from plone import api
 
 
-def cook_css_resources(context):
-    """Cook css resources."""
+def cook_css_resources(context):  # pragma: no cover
+    """Cook CSS resources."""
     css_tool = api.portal.get_tool('portal_css')
     css_tool.cookResources()
     logger.info('CSS resources were cooked')
 
 
-def cook_javascript_resources(context):
-    """Cook javascript resources."""
+def cook_javascript_resources(context):  # pragma: no cover
+    """Cook JavaScript resources."""
     js_tool = api.portal.get_tool('portal_javascripts')
     js_tool.cookResources()
     logger.info('Javascript resources were cooked')